Add 'are you sure' when deleting all rounds

+++ src/app/app.component.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core'
 import { NavigationStart, Router } from '@angular/router'
 import { filter } from 'rxjs'
-import {GameService} from "./game/game.service";
-import {MatSnackBar} from "@angular/material/snack-bar";
-import {Round} from "./game/game.model";
-import {ReorderTeamsComponent} from "./pages/current-round-page/result-input/reorder-teams/reorder-teams.component";
-import {MatDialog} from "@angular/material/dialog";
+import { GameService } from './game/game.service'
+import { MatSnackBar } from '@angular/material/snack-bar'
+import { Round } from './game/game.model'
+import { ReorderTeamsComponent } from './pages/current-round-page/result-input/reorder-teams/reorder-teams.component'
+import { MatDialog } from '@angular/material/dialog'
 
 @Component({
   selector: 'app-root',
@@ -13,8 +13,12 @@ import {MatDialog} from "@angular/material/dialog";
   styleUrls: ['./app.component.css'],
 })
 export class AppComponent {
-
-  constructor(private router: Router, private gameService: GameService, private _snackBar: MatSnackBar, public dialog: MatDialog) {
+  constructor(
+    private router: Router,
+    private gameService: GameService,
+    private _snackBar: MatSnackBar,
+    public dialog: MatDialog
+  ) {
     // @ts-ignore
     router.events.pipe(filter((event) => event instanceof NavigationStart)).subscribe((event: NavigationStart) => {
       this.activeLink = event.url
@@ -45,17 +49,19 @@ export class AppComponent {
   }
 
   deleteAllRounds(): void {
-    this.gameService.deleteAllRounds().subscribe({
-      next: () => {
-        this._snackBar.open('Rounds deleted', undefined, {
-          duration: 3000,
-        })
-      },
-      error: (error) =>
-        this._snackBar.open(error.message, undefined, {
-          duration: 3000,
-        }),
-    })
+    if (confirm('Are you sure to delete all rounds?')) {
+      this.gameService.deleteAllRounds().subscribe({
+        next: () => {
+          this._snackBar.open('Rounds deleted', undefined, {
+            duration: 3000,
+          })
+        },
+        error: (error) =>
+          this._snackBar.open(error.message, undefined, {
+            duration: 3000,
+          }),
+      })
+    }
   }
 
   openReorderTeamsModal(currentRound: Round) {
@@ -63,9 +69,6 @@ export class AppComponent {
       data: currentRound,
     })
 
-    dialogRef.afterClosed().subscribe((result) => {
-      console.log('The dialog was closed')
-      console.log(result)
-    })
+    dialogRef.afterClosed().subscribe(() => {})
   }
 }
